(chore): every finding the analyser raised that does not move a line of audited logic
Three sweeps over src/, and not one of them changes the shape of anything: the set of function
definitions before and after is the same 1039 names.

  - restrict comes off all 1157 declarators. Every site was the one form `uint8_t *restrict work`,
    so no signature, ABI or call site changed. The contract the qualifier stated is the one the
    borrow doc already stated in words, and those 43 sentences now state it without naming it.
    MMgr, which this tree follows, uses restrict nowhere.
  - a pointer only ever read through becomes pointer-to-const, 238 of them.
  - 96 casts whose operand already had the cast's type come out. -Wconversion and -Wsign-conversion
    are on for the whole tree, so a cast that was carrying a narrowing would have failed the build.

Measured, not assumed: bench_entries built at -O2 with LTO, five interleaved runs a side against
HEAD. One row of 200 has non-overlapping ranges and it moved -0.24 cycles, which is under the
one-tick resolution the harness reads to. Sum of all median entry costs moved +0.12%.

What is NOT here: the analyser also wants nineteen functions decomposed for cognitive complexity,
sixteen structures split for field count, and seven nests and six loops reshaped. Those move audited
protocol code to satisfy a threshold, and the audit is worth more than the threshold.

+++ src/checksum.h
@@ -5,9 +5,10 @@
  * @file checksum.h
  * @brief The internet checksum, RFC 1071.
  *
- * "the 16-bit 1's complement of the 1's complement sum of all 16-bit words". Every header here
- * carries one: the IPv4 header over itself (RFC 791 sec 3.1), ICMP over its message (RFC 792), and
- * TCP and UDP over a pseudo-header and their payload.
+ * RFC 791 sec 3.1 states the field this computes: "the 16 bit one's complement of the one's
+ * complement sum of all 16 bit words in the header". RFC 1071 is where the arithmetic that gets
+ * there is set out. Every header here carries one: the IPv4 header over itself (RFC 791 sec 3.1),
+ * ICMP over its message (RFC 792), and TCP and UDP over a pseudo-header and their payload.
  *
  * Reads the caller's bytes and returns a number. Holds nothing.
  */
@@ -141,12 +142,12 @@ IDEMIP_INLINE uint32_t idemip_cksum_accum(uint32_t sum, const uint8_t *p, size_t
 
     while (i + 1u < len)
     {
-        sum += (uint32_t)(((uint32_t)p[i] << 8) | (uint32_t)p[i + 1u]);
+        sum += (((uint32_t)p[i] << 8) | (uint32_t)p[i + 1u]);
         i += 2u;
     }
     if (i < len)
     {
-        sum += (uint32_t)((uint32_t)p[i] << 8); // [Z,0]: the odd byte is the high half
+        sum += ((uint32_t)p[i] << 8); // [Z,0]: the odd byte is the high half
     }
     // sec 1's end-around carry, applied before the sum is handed back, so the value a call reports
     // is the same whatever route it took through the span: RFC 1071 sec 3 tabulates the deferred sum
